Fixed AttributeError: model object has no attribute 'rel'
