+++ datatableview/columns.py
@@ -23,6 +23,7 @@ try:
     from django.utils.encoding import python_2_unicode_compatible
 except ImportError:
     from .compat import python_2_unicode_compatible
+from .compat import get_rel_to
 
 import six
 import dateutil.parser
@@ -47,15 +48,6 @@ STRPTIME_PLACEHOLDERS = {
 def register_simple_modelfield(model_field):
     column_class = get_column_for_modelfield(model_field)
     COLUMN_CLASSES.insert(0, (column_class, [model_field]))
-
-def get_rel_to(model_field):
-    # as per https://gist.github.com/ofalk/404f9f637b7b4520e26dcdd520dbe248
-    rel_to = None
-    if hasattr(model_field, 'rel'):
-        rel_to = model_field.rel.to if model_field.rel else None
-    elif hasattr(model_field, 'remote_field'):
-        rel_to = model_field.remote_field.model if model_field.remote_field else None
-    return rel_to
 
 def get_column_for_modelfield(model_field):
     """ Return the built-in Column class for a model field class. """
